+++ src/shared/Auth/Sha1.cpp
@@ -37,7 +37,7 @@ void Sha1Hash::UpdateData(const uint8 *dta, int len)
 
 void Sha1Hash::UpdateData(const std::string &str)
 {
-	UpdateData((uint8 *)str.c_str(), str.length());
+	UpdateData((uint8 *)str.c_str(), (int)str.length());
 }
 
 void Sha1Hash::UpdateBigNumbers(BigNumber *bn0, ...)
